requirements.txt
Flask==1.1.2
Flask-Cors==3.0.3
Flask-Script==2.0.6
gunicorn==19.7.1
isodate==0.6.0
itsdangerous==0.24
Werkzeug==0.15.5
requests==2.24.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Flask-PyMongo==2.3.0
pymongo==3.10.1
weaviate-client==3.9.0
loguru==0.6.0
Flask-Bcrypt==1.0.1
inject==5.0.0
openai==2.2.0
PyPDF2==3.0.1
python-docx==0.8.11 
openpyxl==3.0.10
pillow==9.5.0
pytesseract==0.3.10
python-pptx==0.6.23
beautifulsoup4==4.12.3
lxml==4.9.4
xlrd==1.2.0
opencv-python==4.12.0.88
numpy==1.24.4
PyMuPDF==1.24.11
python-dotenv==0.21.1
flask-restx==1.1.0